% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burritos_cost.R
\name{burritos_cost}
\alias{burritos_cost}
\title{Calculate the total price of burritos}
\usage{
burritos_cost(price, count)
}
\arguments{
\item{price}{cost of 1burrito}

\item{count}{number of burritos ordered}
}
\description{
Calculate the total price of burritos
}
