% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cm3_m3.R
\name{convert_cm3_m3}
\alias{convert_cm3_m3}
\title{Convert cubic centimeters to cubic meters}
\usage{
convert_cm3_m3(cm3)
}
\arguments{
\item{cm3}{insert value in cubic centimeters}
}
\value{
equivalent value in cubic meters
}
\description{
Convert cubic centimeters to cubic meters
}
